refactor: split grepDisabledCommand out and renamed to grepCommand
