+++ src/index.ts
@@ -1,5 +1,6 @@
 export * from './findDisabled';
 export * from './gitCommand';
+export * from './grepCommand';
 export * from './reportDisabled';
 export * from './scripts';
 export * from './types';
